+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## CHANGELOG
-#2022-05-27 Mobile Version 3.20.0 Ui 3.20.0
-- feat(mobile&ui): Added Origin Verification.
-- 
+
+# 2023-11-23 Mobile Version 3.20.3 Ui 3.17.3
+- feat(mobile&ui): Update default bank text from evolve to Patriot.
+
 # 2022-01-13 Mobile Version 3.17.2 Ui 3.17.2
 - feat(mobile&ui): Minor updates to improve onboarding experience.